refactored and built for deployment
