+++ package.json
@@ -17,6 +17,7 @@
 	"devDependencies": {
 		"@babel/preset-env": "^7.26.0",
 		"@babel/preset-react": "^7.26.3",
+		"@next/bundle-analyzer": "^15.0.3",
 		"@testing-library/dom": "^10.4.0",
 		"@testing-library/jest-dom": "^6.6.3",
 		"@testing-library/react": "^16.0.1",
